feat!: move endpoints from service to serviceRelease
the following endpoints have been moved to servicesRelease
POST: /api/services/addservice
PUT: /api/services/{serviceId}
PUT: /api/services/{serviceId}/submit
PUT: /api/services/{serviceId}/approveService
PUT: /api/services/{serviceId}/declineService
PUT: /api/services/updateservicedoc/{serviceId}/documentType/{documentTypeId}/documents

Refs: CPLP-2475
BREAKING CHANGES: Endpoints have been moved
