+++ src/marketplace/Services.Service/BusinessLogic/IServiceReleaseBusinessLogic.cs
@@ -22,6 +22,7 @@ using Org.Eclipse.TractusX.Portal.Backend.Framework.Models;
 using Org.Eclipse.TractusX.Portal.Backend.PortalBackend.DBAccess.Models;
 using Org.Eclipse.TractusX.Portal.Backend.Services.Service.ViewModels;
 using Org.Eclipse.TractusX.Portal.Backend.Offers.Library.Models;
+using Org.Eclipse.TractusX.Portal.Backend.PortalBackend.PortalEntities.Enums;
 
 namespace Org.Eclipse.TractusX.Portal.Backend.Services.Service.BusinessLogic;
 
@@ -82,6 +83,56 @@ public interface IServiceReleaseBusinessLogic
     /// <param name="serviceName"></param>
     /// <param name="languageShortName"></param>
     Task<Pagination.Response<InReviewServiceData>> GetAllInReviewStatusServiceAsync(int page, int size, OfferSorting? sorting, string? serviceName, string? languageShortName);
+    
+    /// <summary>
+    /// Creates a new service offering
+    /// </summary>
+    /// <param name="data">The data to create the service offering</param>
+    /// <param name="iamUserId">the iamUser id</param>
+    /// <returns>The id of the newly created service</returns>
+    Task<Guid> CreateServiceOfferingAsync(ServiceOfferingData data, string iamUserId);
+
+    /// <summary>
+    /// Updates the given service
+    /// </summary>
+    /// <param name="serviceId">Id of the service to update</param>
+    /// <param name="data">Data of the updated entry</param>
+    /// <param name="iamUserId">Id of the current user</param>
+    Task UpdateServiceAsync(Guid serviceId, ServiceUpdateRequestData data, string iamUserId);
+
+    /// <summary>
+    /// Update app status and create notification
+    /// </summary>
+    /// <param name="serviceId">Id of the service that should be submitted</param>
+    /// <param name="iamUserId">Id of the iamUser</param>
+    Task SubmitServiceAsync(Guid serviceId, string iamUserId);
+
+    /// <summary>
+    /// Approve Service Status from IN_Review to Active
+    /// </summary>
+    /// <param name="appId"></param>
+    /// <param name="iamUserId"></param>
+    /// <returns></returns>
+    Task ApproveServiceRequestAsync(Guid appId, string iamUserId);
+
+    /// <summary>
+    /// Declines the service request
+    /// </summary>
+    /// <param name="serviceId">Id of the service</param>
+    /// <param name="iamUserId">Id of the iamUser</param>
+    /// <param name="data">The decline request data</param>
+    Task DeclineServiceRequestAsync(Guid serviceId, string iamUserId, OfferDeclineRequest data);
+
+    /// <summary>
+    /// Upload document for given company user for Service
+    /// </summary>
+    /// <param name="serviceId"></param>
+    /// <param name="documentTypeId"></param>
+    /// <param name="document"></param>
+    /// <param name="iamUserId"></param>
+    /// <param name="cancellationToken"></param>
+    /// <returns></returns>
+    Task CreateServiceDocumentAsync(Guid serviceId, DocumentTypeId documentTypeId, IFormFile document, string iamUserId, CancellationToken cancellationToken);
 
     /// <summary>
     /// Delete the Service Document
