+++ tests/marketplace/Services.Service.Tests/Controllers/ServiceReleaseControllerTest.cs
@@ -21,6 +21,7 @@
 using AutoFixture;
 using FakeItEasy;
 using FluentAssertions;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Org.Eclipse.TractusX.Portal.Backend.Offers.Library.Models;
 using Org.Eclipse.TractusX.Portal.Backend.Framework.Models;
@@ -29,6 +30,7 @@ using Org.Eclipse.TractusX.Portal.Backend.PortalBackend.PortalEntities.Enums;
 using Org.Eclipse.TractusX.Portal.Backend.Services.Service.BusinessLogic;
 using Org.Eclipse.TractusX.Portal.Backend.Services.Service.Controllers;
 using Org.Eclipse.TractusX.Portal.Backend.Services.Service.ViewModels;
+using Org.Eclipse.TractusX.Portal.Backend.Tests.Shared;
 using Org.Eclipse.TractusX.Portal.Backend.Tests.Shared.Extensions;
 using Xunit;
 
@@ -36,8 +38,9 @@ namespace Org.Eclipse.TractusX.Portal.Backend.Services.Service.Tests.Controllers
 
 public class ServiceReleaseControllerTest
 {
-    private static readonly string IamUserId = "4C1A6851-D4E7-4E10-A011-3732CD045E8A";
-    private readonly string _accessToken = "THISISTHEACCESSTOKEN";
+    private const string IamUserId = "4C1A6851-D4E7-4E10-A011-3732CD045E8A";
+    private const string AccessToken = "THISISTHEACCESSTOKEN";
+    private static readonly Guid ServiceId = new("4C1A6851-D4E7-4E10-A011-3732CD045453");
     private readonly IFixture _fixture;
     private readonly IServiceReleaseBusinessLogic _logic;
     private readonly ServiceReleaseController _controller;
@@ -46,7 +49,7 @@ public class ServiceReleaseControllerTest
         _fixture = new Fixture();
         _logic = A.Fake<IServiceReleaseBusinessLogic>();
         this._controller = new ServiceReleaseController(_logic);
-        _controller.AddControllerContextWithClaimAndBearer(IamUserId, _accessToken);
+        _controller.AddControllerContextWithClaimAndBearer(IamUserId, AccessToken);
     }
 
     [Fact]
@@ -185,5 +188,89 @@ public class ServiceReleaseControllerTest
         Assert.IsType<NoContentResult>(result);
         A.CallTo(() => _logic.DeleteServiceDocumentsAsync(documentId, IamUserId))
             .MustHaveHappenedOnceExactly();
+    }
+    
+    [Fact]
+    public async Task CreateServiceOffering_ReturnsExpectedId()
+    {
+        //Arrange
+        var id = new Guid("d90995fe-1241-4b8d-9f5c-f3909acc6383");
+        var serviceOfferingData = _fixture.Create<ServiceOfferingData>();
+        A.CallTo(() => _logic.CreateServiceOfferingAsync(A<ServiceOfferingData>._, IamUserId))
+            .Returns(id);
+
+        //Act
+        var result = await this._controller.CreateServiceOffering(serviceOfferingData).ConfigureAwait(false);
+
+        //Assert
+        A.CallTo(() => _logic.CreateServiceOfferingAsync(serviceOfferingData, IamUserId)).MustHaveHappenedOnceExactly();
+        Assert.IsType<CreatedAtRouteResult>(result);
+        result.Value.Should().Be(id);
+    }
+ 
+    [Fact]
+    public async Task UpdateService_ReturnsExpected()
+    {
+        //Arrange
+        var serviceId = _fixture.Create<Guid>();
+        var data = _fixture.Create<ServiceUpdateRequestData>();
+        A.CallTo(() => _logic.UpdateServiceAsync(A<Guid>._, A<ServiceUpdateRequestData>._, A<string>.That.Matches(x => x== IamUserId)))
+            .Returns(Task.CompletedTask);
+
+        //Act
+        var result = await this._controller.UpdateService(serviceId, data).ConfigureAwait(false);
+
+        //Assert
+        A.CallTo(() => _logic.UpdateServiceAsync(serviceId, data, IamUserId)).MustHaveHappenedOnceExactly();
+        Assert.IsType<NoContentResult>(result);
+    }
+
+    [Fact]
+    public async Task SubmitService_ReturnsExpectedCount()
+    {
+        //Arrange
+        A.CallTo(() => _logic.SubmitServiceAsync(A<Guid>._, A<string>._))
+            .ReturnsLazily(() => Task.CompletedTask);
+
+        //Act
+        await this._controller.SubmitService(ServiceId).ConfigureAwait(false);
+
+        //Assert
+        A.CallTo(() => _logic.SubmitServiceAsync(ServiceId, IamUserId)).MustHaveHappenedOnceExactly();
+    }
+        
+    [Fact]
+    public async Task DeclineServiceRequest_ReturnsNoContent()
+    {
+        //Arrange
+        var serviceId = _fixture.Create<Guid>();
+        var data = new OfferDeclineRequest("Just a test");
+        A.CallTo(() => _logic.DeclineServiceRequestAsync(A<Guid>._, A<string>._, A<OfferDeclineRequest>._))
+            .ReturnsLazily(() => Task.CompletedTask);
+
+        //Act
+        var result = await this._controller.DeclineServiceRequest(serviceId, data).ConfigureAwait(false);
+
+        //Assert
+        A.CallTo(() => _logic.DeclineServiceRequestAsync(serviceId, IamUserId, data)).MustHaveHappenedOnceExactly();
+        result.Should().BeOfType<NoContentResult>();
+    }
+
+    [Fact]
+    public async Task UpdateServiceDocumentAsync_CallExpected()
+    {
+        // Arrange
+        var serviceId = _fixture.Create<Guid>();
+        var file = FormFileHelper.GetFormFile("this is just a test", "superFile.pdf", "application/pdf");
+        A.CallTo(() => _logic.CreateServiceDocumentAsync(A<Guid>._,
+            A<DocumentTypeId>._, A<IFormFile>._, A<string>._, CancellationToken.None))
+            .ReturnsLazily(() => Task.CompletedTask);
+        
+        // Act
+        await this._controller.UpdateServiceDocumentAsync(serviceId,DocumentTypeId.ADDITIONAL_DETAILS,file,CancellationToken.None).ConfigureAwait(false);
+
+        // Assert
+        A.CallTo(() => _logic.CreateServiceDocumentAsync(serviceId,
+            DocumentTypeId.ADDITIONAL_DETAILS, file, IamUserId, CancellationToken.None)).MustHaveHappened();
     }
 }
